requirements.txt
Django
gunicorn
django-heroku
django-bootstrap4
beautifulsoup4
django-cors-headers
djangorestframework
requests